requirements.txt
discord.py
pillow
ffmpeg
youtube-dl